% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_effsize_ci.R
\name{aov_effsize}
\alias{aov_effsize}
\title{Effect sizes and their CIs for \code{aov} objects}
\usage{
aov_effsize(
  model,
  effsize = "eta",
  partial = TRUE,
  ci = 0.95,
  iterations = 100
)
}
\description{
Effect sizes and their CIs for \code{aov} objects
}
\examples{
\donttest{
# for reproducibility
set.seed(123)

# model
mod <- stats::lm(formula = wt ~ am * cyl, data = mtcars)

# effect size
statsExpressions:::aov_effsize(mod, effsize = "eta")
statsExpressions:::aov_effsize(mod, effsize = "eta", partial = FALSE)
statsExpressions:::aov_effsize(mod, effsize = "omega")
statsExpressions:::aov_effsize(mod, effsize = "omega", partial = FALSE)
}
}
\keyword{internal}
