% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bugs_long}
\alias{bugs_long}
\title{Tidy version of the "Bugs" dataset.}
\format{
A data frame with 372 rows and 6 variables
\itemize{
\item subject. Dummy identity number for each participant.
\item gender. Participant's gender (Female, Male).
\item region. Region of the world the participant was from.
\item education. Level of education.
\item condition. Condition of the experiment the participant gave rating
for (\strong{LDLF}: low disgustingness and low frighteningness; \strong{LDHF}: low
disgustingness and high frighteningness; \strong{HDLF}: high disgustingness
and low frighteningness; \strong{HDHF}: high disgustingness and high
frighteningness).
\item desire. The desire to kill an arthropod was indicated on a scale from
0 to 10.
}
}
\usage{
bugs_long
}
\description{
Tidy version of the "Bugs" dataset.
}
\details{
This data set, "Bugs", provides the extent to which men and women
want to kill arthropods that vary in frighteningness (low, high) and
disgustingness (low, high). Each participant rates their attitudes towards
all arthropods. Subset of the data reported by Ryan et al. (2013).
}
\examples{
dim(bugs_long)
head(bugs_long)
dplyr::glimpse(bugs_long)
}
\references{
Ryan, R. S., Wilde, M., & Crist, S. (2013). Compared to a small, supervised
lab experiment, a large, unsupervised web-based experiment on a previously
unknown effect has benefits that outweigh its potential costs. \emph{Computers in
Human Behavior}, \emph{29}(4), 1295-1301.
}
\keyword{datasets}
