% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_t_test.R
\name{expr_t_bayes}
\alias{expr_t_bayes}
\title{Making expression containing Bayesian \emph{t}-test results}
\usage{
expr_t_bayes(
  data,
  x,
  y,
  subject.id = NULL,
  paired = FALSE,
  k = 2L,
  bf.prior = 0.707,
  output = "expression",
  ...
)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{x}{The grouping variable from the dataframe \code{data}.}

\item{y}{The response (a.k.a. outcome or dependent) variable from the
dataframe \code{data}.}

\item{subject.id}{In case of repeated measures design (\code{paired = TRUE},
i.e.), this argument specifies the subject or repeated measures id. Note
that if this argument is \code{NULL} (which is the default), the function
assumes that the data has already been sorted by such an id by the user and
creates an internal identifier. So if your data is \strong{not} sorted and you
leave this argument unspecified, the results can be inaccurate.}

\item{paired}{Logical that decides whether the experimental design is
repeated measures/within-subjects or between-subjects. The default is
\code{FALSE}.}

\item{k}{Number of digits after decimal point (should be an integer)
(Default: \code{k = 2L}).}

\item{bf.prior}{A number between \code{0.5} and \code{2} (default \code{0.707}), the prior
width to use in calculating Bayes factors.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Making expression containing Bayesian \emph{t}-test results
}
\examples{
# for reproducibility
set.seed(123)
library(statsExpressions)

# ------------- between-subjects design --------------------------

expr_t_bayes(
  data = mtcars,
  x = am,
  y = wt,
  paired = FALSE
)

# ------------- within-subjects design -----------------------------

expr_t_bayes(
  data = dplyr::filter(bugs_long, condition \%in\% c("LDLF", "LDHF")),
  x = condition,
  y = desire,
  paired = TRUE,
  subject.id = subject
)
}
\references{
For more details, see-
\url{https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html}
}
