% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_markdown.R
\name{format_markdown}
\alias{format_markdown}
\title{Format expression to R Markdown}
\usage{
format_markdown(expr, ...)
}
\arguments{
\item{expr}{The expression returned inside the data.frame containing
statistical details}

\item{...}{Currently ignored}
}
\description{
Pass the expression generated using any of the functions to
get statistical details ready for the writing reports in R Markdown
}
\examples{
\donttest{
# for reproducibility
set.seed(123)
library(statsExpressions)
options(tibble.width = Inf, pillar.bold = TRUE, pillar.neg = TRUE)

# without changing defaults
result <- corr_test(
  data = ggplot2::midwest,
  x    = area,
  y    = percblack
)

format_markdown(result$expression)

result <- oneway_anova(
  data = iris_long,
  x = condition,
  y = value,
  paired = TRUE
)

format_markdown(result$expression)

}
}
