% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_comparisons.R
\name{p_adjust_text}
\alias{p_adjust_text}
\title{\emph{p}-value adjustment method text}
\usage{
p_adjust_text(p.adjust.method)
}
\arguments{
\item{p.adjust.method}{Adjustment method for \emph{p}-values for multiple
comparisons. Possible methods are: \code{"holm"} (default), \code{"hochberg"},
\code{"hommel"}, \code{"bonferroni"}, \code{"BH"}, \code{"BY"}, \code{"fdr"}, \code{"none"}.}
}
\value{
Standardized text description for what method was used.
}
\description{
Preparing text to describe which \emph{p}-value adjustment method was used
}
\examples{
p_adjust_text("none")
p_adjust_text("BY")

}
