% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_messages.R
\name{stats_type_switch}
\alias{stats_type_switch}
\title{Switch function to determine which type of statistics is to be run.}
\usage{
stats_type_switch(stats.type)
}
\arguments{
\item{stats.type}{Character describing the needed type of statistics (e.g.,
\code{"parametric"}, \code{"nonparametric"}, \code{"robust"}, `"bayes"``, etc.).}
}
\description{
Takes in all allowed characters describing the needed type of
test and converts it into standard terms to reduce the complexity of
conditional statements.
}
\author{
\href{https://github.com/IndrajeetPatil}{Indrajeet Patil}
}
\keyword{internal}
