% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_easystats.R
\name{tidy_model_effectsize}
\alias{tidy_model_effectsize}
\title{Convert \code{effectsize} package output to \code{tidyverse} conventions}
\usage{
tidy_model_effectsize(data)
}
\arguments{
\item{data}{Dataframe returned by \code{effectsize} functions.}
}
\description{
Convert \code{effectsize} package output to \code{tidyverse} conventions
}
\examples{
df <- effectsize::cohens_d(sleep$extra, sleep$group)
tidy_model_effectsize(df)
}
