% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_anova.R
\name{expr_anova_bayes}
\alias{expr_anova_bayes}
\title{Making expression containing Bayesian one-way ANOVA results.}
\usage{
expr_anova_bayes(
  data,
  x,
  y,
  paired = FALSE,
  bf.prior = 0.707,
  k = 2L,
  output = "expression",
  ...
)
}
\arguments{
\item{data}{A dataframe (or a tibble) from which variables specified are to
be taken. A matrix or tables will \strong{not} be accepted.}

\item{x}{The grouping variable from the dataframe \code{data}.}

\item{y}{The response (a.k.a. outcome or dependent) variable from the
dataframe \code{data}.}

\item{paired}{Logical that decides whether the experimental design is
repeated measures/within-subjects or between-subjects. The default is
\code{FALSE}.}

\item{bf.prior}{A number between \code{0.5} and \code{2} (default \code{0.707}), the prior
width to use in calculating Bayes factors.}

\item{k}{Number of digits after decimal point (should be an integer)
(Default: \code{k = 2}).}

\item{output}{Can either be \code{"expression"}, which will return expression or
\code{"dataframe"}, which will return a dataframe with results.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
For more details, see-
\url{https://indrajeetpatil.github.io/statsExpressions/articles/stats_details.html}
}
\description{
Making expression containing Bayesian one-way ANOVA results.
}
\examples{
# setup
set.seed(123)
library(statsExpressions)

# between-subjects ---------------------------------------

expr_anova_bayes(
  data = ggplot2::msleep,
  x = vore,
  y = sleep_rem
)

# repeated measures ---------------------------------------

expr_anova_bayes(
  data = WRS2::WineTasting,
  x = Wine,
  y = Taste,
  paired = TRUE,
  k = 4
)
}
