% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{movies_long}
\alias{movies_long}
\title{Movie information and user ratings from IMDB.}
\format{
A data frame with 1,579 rows and 8 variables
\itemize{
\item title.  Title of the movie.
\item year.  Year of release.
\item budget.  Total budget (if known) in US dollars
\item length.  Length in minutes.
\item rating.  Average IMDB user rating.
\item votes.  Number of IMDB users who rated this movie.
\item mpaa.  MPAA rating.
\item genre. Genre of the movie (\code{"Action"}, \code{"Action Comedy"},
\code{"Action Drama"}, \code{"Animated"}, \code{"Comedy"}, \code{"Comedy Drama"}, \code{"Drama"},
\code{"Romance Drama"}, \code{"RomCom"}).
}
}
\source{
\url{https://CRAN.R-project.org/package=ggplot2movies}
}
\usage{
movies_long
}
\description{
Movie information and user ratings from IMDB.
}
\details{
Modified dataset from \code{{ggplot2movies}} package.
}
\examples{
dim(movies_long)
head(movies_long)
dplyr::glimpse(movies_long)
}
\keyword{datasets}
