% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%<>\%}
\alias{\%$\%}
\alias{\%<-\%}
\alias{\%||\%}
\alias{\%|\%}
\alias{:=}
\alias{tibble}
\alias{as_tibble}
\alias{bf_contingency_tab}
\alias{bf_ttest}
\alias{bf_meta_random}
\alias{bf_corr_test}
\alias{bf_oneway_anova}
\alias{prior}
\alias{correlation}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{correlation}{\code{\link[correlation]{correlation}}}

  \item{ipmisc}{\code{\link[ipmisc:reexports]{\%$\%}}, \code{\link[ipmisc:reexports]{\%<-\%}}, \code{\link[ipmisc:reexports]{\%<>\%}}, \code{\link[ipmisc:reexports]{\%>\%}}, \code{\link[ipmisc:reexports]{as_tibble}}, \code{\link[ipmisc:reexports]{tibble}}}

  \item{rlang}{\code{\link[rlang:op-na-default]{\%|\%}}, \code{\link[rlang:op-null-default]{\%||\%}}, \code{\link[rlang:nse-force]{:=}}}

  \item{tidyBF}{\code{\link[tidyBF]{bf_contingency_tab}}, \code{\link[tidyBF]{bf_corr_test}}, \code{\link[tidyBF]{bf_meta_random}}, \code{\link[tidyBF]{bf_oneway_anova}}, \code{\link[tidyBF]{bf_ttest}}, \code{\link[tidyBF:reexports]{prior}}}
}}

