% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsExpressions-package.R
\docType{package}
\name{statsExpressions-package}
\alias{statsExpressions-package}
\alias{_PACKAGE}
\alias{statsExpressions}
\title{statsExpressions: Tidy Dataframes and Expressions with Statistical Details}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}

\code{statsExpressions} package produces tidy dataframes with rich details for the
most common types of statistical approaches and tests: parametric,
nonparametric, robust, and Bayesian t-test, one-way ANOVA, correlation
analyses, contingency table analyses, and meta-analyses. The functions are
pipe-friendly and provide a consistent syntax to work with tidy data. These
dataframes additionally contain expressions with statistical details, and can
be used in graphing packages. This package also forms the statistical
processing backend for
\href{https://indrajeetpatil.github.io/ggstatsplot/}{\code{ggstatsplot}} package.

For more documentation, see the dedicated
\href{https://indrajeetpatil.github.io/statsExpressions/}{Website}.
}
\details{
\code{statsExpressions}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://indrajeetpatil.github.io/statsExpressions/}
  \item \url{https://github.com/IndrajeetPatil/statsExpressions}
  \item Report bugs at \url{https://github.com/IndrajeetPatil/statsExpressions/issues}
}

}
\author{
\strong{Maintainer}: Indrajeet Patil \email{patilindrajeet.science@gmail.com} (\href{https://orcid.org/0000-0003-1995-6531}{ORCID}) (@patilindrajeets) [copyright holder]

}
